Rename `apps` to `sciapps` in test library
