+++ hpctestlib/sciapps/jupyter/ipcmagic/__init__.py
