+++ hpctestlib/sciapps/jupyter/ipcmagic/src/tf-hvd-sgd-ipc-tf2.py
